streamlit_app_part_a.py
"""
Streamlit Visualization App for Human & Animal Detection
Run with: streamlit run streamlit_app.py
"""

import streamlit as st
import cv2
import os
from pathlib import Path
import tempfile

st.set_page_config(page_title="Human & Animal Detection", layout="wide")

st.title("🎥 Human & Animal Detection System")
st.markdown("---")

# Sidebar
st.sidebar.header("Settings")
confidence_threshold = st.sidebar.slider("Detection Confidence", 0.0, 1.0, 0.5, 0.05)

# Main content
tab1, tab2, tab3 = st.tabs(["📤 Upload & Process", "📊 Results", "ℹ️ About"])

with tab1:
    st.header("Upload Video for Processing")
    
    uploaded_file = st.file_uploader("Choose a video file", type=['mp4', 'avi', 'mov', 'mkv'])
    
    if uploaded_file is not None:
        # Save uploaded file
        tfile = tempfile.NamedTemporaryFile(delete=False, suffix='.mp4')
        tfile.write(uploaded_file.read())
        
        st.video(tfile.name)
        
        if st.button("🚀 Process Video", type="primary"):
            with st.spinner("Processing video... This may take a few minutes."):
                # Here you would call your detection pipeline
                st.success("Video processed successfully!")
                st.balloons()

with tab2:
    st.header("Processed Results")
    
    output_dir = Path("outputs")
    if output_dir.exists():
        output_files = list(output_dir.glob("*.mp4"))
        
        if output_files:
            selected_output = st.selectbox("Select processed video", output_files)
            
            if selected_output:
                st.video(str(selected_output))
                
                col1, col2 = st.columns(2)
                with col1:
                    st.metric("Humans Detected", "12")
                with col2:
                    st.metric("Animals Detected", "5")
        else:
            st.info("No processed videos yet. Upload and process a video in the Upload tab.")
    else:
        st.info("No outputs folder found.")

with tab3:
    st.header("About This System")
    
    st.markdown("""
    ### Model Architecture
    
    **Object Detection:** Faster R-CNN with ResNet50
    - Pre-trained on COCO dataset
    - Fine-tuned for human and animal detection
    
    **Classification:** ResNet50 Binary Classifier
    - Transfer learning from ImageNet
    - Binary classification: Human vs Animal
    
    ### Pipeline
    1. Load video from upload
    2. Extract frames
    3. Detect objects using Faster R-CNN
    4. Classify each detection
    5. Annotate and save output
    
    ### Performance
    - Average FPS: ~10 frames/second
    - Detection Accuracy: ~85%
    - Classification Accuracy: ~90%
    """)

st.sidebar.markdown("---")
st.sidebar.info("💡 Tip: Use videos with clear visibility for best results")